check if its admin user or managr and delete the user

+++ culina/controllers/CompanyUser.js
@@ -77,18 +77,22 @@ exports.removeCompanyUser = async function (
   res
 ) {
   try {
-    await CompanyUser.findOneAndRemove({ _id: userId });
-    const company = await Company.findById({ _id: companyId });
+    // NOTE: only amin user can remove users and Manager user
+    // Check if admin or manager
+    if (admin || (position === "Manager" && company === companyId)) {
+      await CompanyUser.findOneAndRemove({ _id: userId });
+      const company = await Company.findById({ _id: companyId });
 
-    if (!company) {
-      return res.status(404).json({ msg: "Company does not exist" });
+      if (!company) {
+        return res.status(404).json({ msg: "Company does not exist" });
+      }
+
+      company.users = company.users.filter((user) => user.userId != userId);
+
+      await company.save();
+
+      res.json({ msg: "USER deleted" });
     }
-
-    company.users = company.users.filter((user) => user.userId != userId);
-
-    await company.save();
-
-    res.json({ msg: "USER deleted" });
   } catch (err) {
     console.error(err);
     res.status(500).send({ error: "Server error" });
